Changed type of ticket form ID from Integer to Long. (#183)

+++ src/test/java/org/zendesk/client/v2/RealSmokeTest.java
@@ -156,7 +156,7 @@ public class RealSmokeTest {
     public void getTicketFieldsOnForm() throws Exception {
         createClientWithTokenOrPassword();
         TicketForm ticketForm = instance.getTicketForm(27562);
-        for(Integer id :ticketForm.getTicketFieldIds()){
+        for(Long id :ticketForm.getTicketFieldIds()){
             Field f = instance.getTicketField(id);
             assertNotNull(f);
         }
